requirements.txt
streamlit
pandas
langchain
langchain-openai
langchain-google-genai
langchain-anthropic
python-dotenv
requests
pygments
